requirements.txt
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
cryptography==3.4.8
gTTS==2.2.3
idna==3.2
pdfminer.six==20201018
pycparser==2.20
requests==2.26.0
six==1.16.0
sortedcontainers==2.4.0
urllib3==1.26.7
